pyproject.toml
[tool.poetry]
name = "dj-asgi-utils"
version = "0.1.0"
description = ""
authors = ["Joshua Massover <massover@simplebet.io>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.1"
daphne = "^2.5.0"
psycopg2 = "^2.8.5"
djangorestframework = "^3.11.1"

[tool.poetry.dev-dependencies]
django-extensions = "^3.0.5"
pytest-django = "^3.9.0"
black = "^19.10b0"
isort = "^5.4.2"
django-bootstrap4 = "^2.2.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
